Created dedicated testing module for radial_pvd_vs_r
